+++ src/lib/validator/valPlainObject.js
@@ -79,10 +79,11 @@ export const validate = function valPlainObject(
       impl.toTypeset(type, q, args),
       [],
       impl.toTypeset(type, q, args, true),
-      extraProps &&
-        new Error(
-          `Found unexpected properties in value: [${extraProps.join(', ')}]`
-        )
+      extraProps.length > 0
+        ? new Error(
+            `Found unexpected properties in value: '${extraProps.join("', '")}'`
+          )
+        : undefined
     );
   }